pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.3.12",
    "livekit-plugins-silero>=1.3.12",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
]
